updated structure of the SearchBar so that it loads in the Search, Filter, and Sort components
