+++ money/src/CardContainer.js
@@ -11,7 +11,7 @@ function CardContainer({ coinData }) {
 
   return (
     <Container>
-      <Row xs={1} md={2} className="g-4">
+      <Row xs={1} md={2} className="g-4 mt-1">
         {deck}
       </Row>
     </Container>
